Add files via upload
The lor_96.csv file contains the data for the full problem, a 100-dimensional Lorentz 96 model modified so that one of the nonlinear terms is small, with noise added. I wrote the data to the csv by putting it in a pandas dataframe, so that may be the easiest way to read it back out. If you do, then the index will be the timestep and the column names are just the numbers 0-99, indexing the components of the model. The file lorentz_filtering_task.py shows how the data is generated. The functions used can be found in lorentz_96.py. The most important is step_particles, which takes in a list of particles and the length of the desired timestep and returns the evolved particle positions. It has an optional argument to set the dimension of the problem, but the default is 100, which as mentioned above is the dimension I chose.